// src/app/shared/enums/bot-state.ts
export enum BotState {
  MENU_INICIAL = 'MENU_INICIAL',
  SOU_CLIENTE = 'SOU_CLIENTE',
  FINANCEIRO = 'FINANCEIRO',
  AGUARDA_CPF_FATURA = 'AGUARDA_CPF_FATURA',
  AGUARDA_CPF_DESBLOQUEIO = 'AGUARDA_CPF_DESBLOQUEIO',
  CONFIRMA_IDENTIDADE_FATURA = 'CONFIRMA_IDENTIDADE_FATURA',
  CONFIRMA_IDENTIDADE_DESBLOQUEIO = 'CONFIRMA_IDENTIDADE_DESBLOQUEIO',
  TRANSFERIDO = 'TRANSFERIDO',
  ENCERRADO = 'ENCERRADO'
}